Support for installing with Puppet 4
This patch adds support to spin up a VM with Puppet 4. Examples are in
the new base boxes, named centos{6,7}-{capsule,katello}-p4-nightly.

This patch has my name on it due to a squash, but much of the credit
goes to @stbenjam and others who made changes, commented, and reviewed
this patch.

NOTE: there is no foreman-installer option for selecting Puppet 3 or
Puppet 4. For a fresh install, the installer will use whichever version
of Puppet is available via the Yum repos configured on the machine.