Refactor box loading by rewriting version parsing
This refactors versions to a list of installers. They know the foreman
version and katello versions and the supported base boxes. This allows
us to note that centos 6 is no longer supported on 1.13 and nightly.

It should also open the path to supporting various Fedora versions and
other distros like Debian.

Closes GH-303