run fb-proxy.bats on the katello box
fb-proxy.bats needs working org setup and actual content to work.
and it also needs working hammer. all that is present on the katello box
during the bats run, but not afterwards. so lets just run fb-proxy.bats
also during the bats run on the katello box.

not including it into the default tests as it needs an additional box to
be set up, which we do not have during *all* runs.