Cert for webpack on chrome (#662)
I was having an issue where it seems JS wasn't loading and UI elements could not be used. It seemed to be caused by ERR_CERT_AUTHORITY_INVALID issues. 

This was running chrome, so I didn't try this at first.

I followed these instruction and it seems to have fixed my issue.