test pipeline for a loadbalanced proxy setup (#684)
This squash commit creates a basic LB smartproxy setup with one Katello, two smartproxies, one HAProxy load balancer in front of the smartproxies, and one client connected to haproxy. It uses 18GB of mem.

Note that the client is already registered via the haproxy via `bootstrap.py` at the end of this setup.