enable CentOS CR repository when using EPEL
EPEL is sometimes compiled against packages that are not in a released
CentOS version yet (because it was built against staging/RHEL). The CR
repository should contain exactly those pre-release bits.

Fixes the issue we see today with rubygem-rkerberos that requires a
newer libkrb5 than in CentOS 7.4

Not using the `yum_repository` module, because it is not able to enable
repositories without defining all the details of the repository, which
seems backwards, as the repo file is already present.